APIs/httpbin.org/0.9.2/python-aiohttp-client/openapi_server/controllers/status_codes_controller.py
from typing import List, Dict
from aiohttp import web

from openapi_server import util


async def status_codes_delete(request: web.Request, codes) -> web.Response:
    """Return status code or random status code if more than one are given

    

    :param codes: 
    :type codes: str

    """
    return web.Response(status=200)


async def status_codes_get(request: web.Request, codes) -> web.Response:
    """Return status code or random status code if more than one are given

    

    :param codes: 
    :type codes: str

    """
    return web.Response(status=200)


async def status_codes_patch(request: web.Request, codes) -> web.Response:
    """Return status code or random status code if more than one are given

    

    :param codes: 
    :type codes: str

    """
    return web.Response(status=200)


async def status_codes_post(request: web.Request, codes) -> web.Response:
    """Return status code or random status code if more than one are given

    

    :param codes: 
    :type codes: str

    """
    return web.Response(status=200)


async def status_codes_put(request: web.Request, codes) -> web.Response:
    """Return status code or random status code if more than one are given

    

    :param codes: 
    :type codes: str

    """
    return web.Response(status=200)


async def status_codes_trace(request: web.Request, codes) -> web.Response:
    """Return status code or random status code if more than one are given

    

    :param codes: 
    :type codes: str

    """
    return web.Response(status=200)
